% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{Chris's log_summed_exps function}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{A vector of type numeric}
}
\value{
The log of the sum of the elements exponentiated.
}
\description{
Chris's log_summed_exps function
}
\examples{
log_summed_exps(1:2000)
log_summed_exps(20)

}
\author{
Chris Chen
}
